requirements.txt
streamlit
pymupdf
openai
pillow
python-dotenv
